Adds support for the date parameter.
This makes it possible to specify a date to use when generating referrals. It also makes it possible to dynamically generate a set of dates based on future dates, past dates, and dates starting on Jan 1st.

This is implemented to make it easier to auto-generate dates for debugging purposes.